Initial work on a Character Sheet output
Involved refactoring the global skills list to an Observeable collection that could be subscribed to by various components.
Note it does have a bug in that the list is never cleared when a skill is overriden - probable fix is to use an ObserveableDictionary so each new row puts its SkillRecord into a very particular place.

There's a lot of manal work involved in creating the SkillRecords and their various subclasses. This is obviously very brittle. Need to consider a way to migrate it all into the JSON somehow...